extract/requirements.txt
tweepy
configparser
requests
mysql-connector-python
schedule
pytz
pandas